// src/findjobnu-api/models/Skill.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FindjobnuService
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SkillProficiency } from './SkillProficiency';
import {
    SkillProficiencyFromJSON,
    SkillProficiencyFromJSONTyped,
    SkillProficiencyToJSON,
    SkillProficiencyToJSONTyped,
} from './SkillProficiency';
import type { Profile } from './Profile';
import {
    ProfileFromJSON,
    ProfileFromJSONTyped,
    ProfileToJSON,
    ProfileToJSONTyped,
} from './Profile';

/**
 * 
 * @export
 * @interface Skill
 */
export interface Skill {
    /**
     * 
     * @type {number}
     * @memberof Skill
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Skill
     */
    name: string;
    /**
     * 
     * @type {SkillProficiency}
     * @memberof Skill
     */
    proficiency: SkillProficiency;
    /**
     * 
     * @type {number}
     * @memberof Skill
     */
    profileId?: number;
    /**
     * 
     * @type {Profile}
     * @memberof Skill
     */
    profile?: Profile;
}



/**
 * Check if a given object implements the Skill interface.
 */
export function instanceOfSkill(value: object): value is Skill {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('proficiency' in value) || value['proficiency'] === undefined) return false;
    return true;
}

export function SkillFromJSON(json: any): Skill {
    return SkillFromJSONTyped(json, false);
}

export function SkillFromJSONTyped(json: any, ignoreDiscriminator: boolean): Skill {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'],
        'proficiency': SkillProficiencyFromJSON(json['proficiency']),
        'profileId': json['profileId'] == null ? undefined : json['profileId'],
        'profile': json['profile'] == null ? undefined : ProfileFromJSON(json['profile']),
    };
}

export function SkillToJSON(json: any): Skill {
    return SkillToJSONTyped(json, false);
}

export function SkillToJSONTyped(value?: Skill | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'proficiency': SkillProficiencyToJSON(value['proficiency']),
        'profileId': value['profileId'],
        'profile': ProfileToJSON(value['profile']),
    };
}